Projekt/requirements.txt
dash==2.14.1
pandas
matplotlib.pyplot
seaborn
plotly.express
dash_bootstrap_components
hashlib
dash_bootstrap_templates
